requirements.txt
Flask==2.3.3
newspaper3k==0.2.8